feat: strategy service consolide methods segregated
