+++ app/adapters/controllers/gameplay/controller.py
@@ -4,9 +4,6 @@ from app.infraestructure.logging import logger
 
 
 class GameplayController:
-    STRATEGY_PATH = 'gameplay/strategies'
-    SEARCH_PAHTERN = f"{STRATEGY_PATH}/*.py"
-
     def __init__(self, starts='welcome', context: dict = None):
         self.finished = False
         self.next = starts
@@ -61,37 +58,3 @@ class GameplayController:
         scenario(self.context, *self.args)
         self.screen.show(scenario.template, self.context)
         self.finished = self.context.get('finished', False)
-
-    def load_strategies(self) -> dict[str, callable]:
-        from importlib import import_module
-        from inspect import getmembers, isfunction, ismethod, getmodule
-
-        strategy_modules = self._map_strategy_modules()
-        functions = {}
-        for mod_path in strategy_modules:
-            module = import_module(mod_path)
-            # get all functions of module
-            found_functions = {
-                **dict(getmembers(module, isfunction)),
-                # support modular decorated strategies
-                **dict(getmembers(module, ismethod))
-            }
-
-            # storage functions
-            functions.update({
-                name: func for name, func in found_functions.items()
-                # when function from the same package of module
-                if getmodule(func).__package__ == module.__package__
-            })
-        logger.debug(f"Loaded strategies: {functions}")
-        return functions
-
-    @classmethod
-    def _map_strategy_modules(cls) -> List[str]:
-        from glob import glob
-        return [
-            file.replace(
-                f"{cls.STRATEGY_PATH}", cls.STRATEGY_PATH
-            ).replace('/', '.')[:-3]
-            for file in glob(cls.SEARCH_PAHTERN)
-        ]
