+++ app/core/domain/services/__init__.py
@@ -1,4 +1,5 @@
 from .planet import PlanetService
+from .strategies import StrategyService
 from .civilization import CivilizationService
 from .skirmish import SkirmishService
 from .memories import MemoriesServiceWrapper
@@ -7,6 +8,7 @@ from .round import RoundService
 
 all = [
     PlanetService,
+    StrategyService,
     CivilizationService,
     SkirmishService,
     MemoriesServiceWrapper,
